Move functionality to make URL unique from UrlRewriteObserver to UrlKeyObserver
